Make a number of cosmetic adjustments [ci skip]

+++ src/format/mod.rs
@@ -7,7 +7,7 @@ trait Validate {
 
 macro_rules! buffer(
     ($capacity:expr) => ({
-        let capacity = $capacity;
+        let capacity = $capacity as usize;
         let mut buffer = Vec::with_capacity(capacity);
         buffer.set_len(capacity);
         buffer
